bugfix(DPAV-1399): fix uncaught exception from invalid geometry (#379)
* bugfix(DPAV-1399): fix uncaught exception from invalid geometry

Couple of issues was causing this, a LineString was returning Position[] which
caused an error from multiLineString. This is now converted in createSegmentCoords.

Care homes were being treated as linear assets due to it having geomtry set (but of
type Point).

- Wrap assets filtering in try/catch to prevent page blowing up
- Update createSegmentCoords to always return a Position[][]
- Fix Asset.isPointAsset check to check geometry type as well

* bugfix: remove typescript-plugin-css-modules to avoid removed package stylus

Removing as this is only a dev dependency and it doesn't seem to be
added as a plugin anyway:
https://www.npmjs.com/package/typescript-plugin-css-modules

See: https://github.com/stylus/stylus/issues/2938

* bugfix: improve asset filtering in polygon

- Prefer checking with lat/lon if set otherwise check with booleanIntersects.
- Remove now unused getters.