feat(DPAV-1503): Add pharmaceutical assets from NHSBSA ODP (#391)
Adds a script to fetch the latest Pharmaceutical List data from the
NHS Business Services Authority Open Data Portal. This is only updated
quarterly so it isn't fetched in real time.

Uses Photon API from OpenStretMap for geocoding as the data does not
provide coordinates only street addresses.

As this will be updated infrequently the script just dumps out the
pharmacies as asset json and it's expected the data file will be
manually updated.